🛠️ Refactor: Extract graph logic from routes to lib
- **Problem:** `src/routes/graph.ts` contained shared domain logic. Also, `src/routes/separation_degrees/[from]/[to]/+page.server.ts` had a lot of business logic to fetch shortest paths.
- **Solution:**
  1. Positioned `graph.ts` into `src/lib/graph.ts` (Colocation, proper structure).
  2. Extracted `getShortestPath` from `+page.server.ts` into `$lib/graph.ts` (Single Responsibility Principle, reducing complexity in endpoints).
  3. Replaced missing `error` parameters with a centralized error handling integration `reportError`.
- **References:** _Clean Code_ (Martin) - Single Responsibility Principle; _Refactoring_ (Fowler) - Extract Method.
- **Verification:** Verified building correctly and passed all checks.
