+++ src/lib/graph.ts
@@ -54,6 +54,38 @@ export const verticesDestino = [
 	...new Set(verticesOrigem.map((origem) => Object.keys(vertices[origem] || {})).flat())
 ].sort();
 
+export function getShortestPath(from: string, to: string) {
+	const mkKey = (key: string) => `Vértices/${key}`;
+	const fromKey = mkKey(from);
+	const toKey = mkKey(to);
+	if (fromKey === toKey) {
+		return null;
+	}
+	const paths: Record<string, string[] | undefined> = {};
+	const visit: Record<string, boolean> = {};
+	paths[fromKey] = [];
+	const queue: string[] = []; // como todos tem o mesmo peso dá pra reduzir o overhead com uma pinha
+	queue.push(fromKey);
+	while (queue.length > 0) {
+		const currentNode = queue.pop() as string; // pop de vetor não vazio não entrega undefined
+		for (const successor of Object.keys(vertices[currentNode] || {})) {
+			const proposition = [...(paths[currentNode] || []), currentNode];
+			if (!paths[successor] || (paths[successor]?.length ?? Infinity) > proposition.length) {
+				paths[successor] = proposition;
+			}
+			if (!visit[successor]) {
+				queue.push(successor);
+				visit[successor] = true;
+			}
+		}
+	}
+	const finalPath = paths[toKey];
+	if (!finalPath) {
+		return null;
+	}
+	return [...finalPath, toKey];
+}
+
 export const db = (async function () {
 	const db = await create({
 		schema: {
