+++ src/routes/separation_degrees/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { Button, Container } from '@sveltestrap/sveltestrap';
 	import VerticeSearch from '../../components/VerticeSearch.svelte';
-	import { typedData, type ObsidianRecord } from '../graph';
+	import { typedData, type ObsidianRecord } from '$lib/graph';
 	import { goto } from '$app/navigation';
 	let originKey = '';
 	let destinationKey = '';
